+++ src/main/java/org/example/hwshop/controller/StorageController.java
@@ -1,8 +1,9 @@
-package org.example.hwshop.domain;
+package org.example.hwshop.controller;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
+import org.example.hwshop.domain.Hardware;
 
 @Data
 @NoArgsConstructor