+++ src/main/java/org/example/hwshop/dto/DiskDTO.java
@@ -1,0 +1,14 @@
+package org.example.hwshop.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class DiskDTO {
+    private long id;
+    private String manufacturer;
+    private String model;
+    private String size;
+    private String port;
+}