+++ src/main/java/org/example/hwshop/service/SetupRepository.java
@@ -1,4 +1,0 @@
-package org.example.hwshop.service;
-
-public class SetupRepository {
-}